+++ fpdf/drawing.py
@@ -3169,7 +3169,6 @@ class DrawingContext:
             return ""
 
         style_dict_name = gsd_registry.register_style(style)
-
         if style_dict_name is not None:
             render_list.insert(2, f"{render_pdf_primitive(style_dict_name)} gs")
             render_list.insert(
@@ -3234,7 +3233,6 @@ class DrawingContext:
                 return ""
 
             style_dict_name = gsd_registry.register_style(style)
-
             if style_dict_name is not None:
                 render_list.insert(2, f"{render_pdf_primitive(style_dict_name)} gs")
                 render_list.insert(