+++ test/image/test_oversized.py
@@ -27,9 +27,9 @@ def test_oversized_images_downscale_simple(caplog, tmp_path):
     pdf.add_page()
     pdf.image(IMAGE_PATH, w=50)
     assert "OVERSIZED: Generated new low-res image" in caplog.text
-    assert len(pdf.images) == 2, pdf.images.keys()
+    assert len(pdf.image_cache.images) == 2, pdf.image_cache.images.keys()
     in_use_img_names = _in_use_img_names(pdf)
-    assert len(in_use_img_names) == 1, pdf.images.keys()
+    assert len(in_use_img_names) == 1, pdf.image_cache.images.keys()
     assert_pdf_equal(pdf, HERE / "oversized_images_downscale_simple.pdf", tmp_path)
 
 
@@ -40,9 +40,9 @@ def test_oversized_images_downscale_twice(tmp_path):
     pdf.add_page()
     pdf.image(IMAGE_PATH, w=50)
     pdf.image(IMAGE_PATH, w=50)
-    assert len(pdf.images) == 2, pdf.images.keys()
+    assert len(pdf.image_cache.images) == 2, pdf.image_cache.images.keys()
     in_use_img_names = _in_use_img_names(pdf)
-    assert len(in_use_img_names) == 1, pdf.images.keys()
+    assert len(in_use_img_names) == 1, pdf.image_cache.images.keys()
     assert_pdf_equal(pdf, HERE / "oversized_images_downscale_twice.pdf", tmp_path)
 
 
@@ -53,7 +53,7 @@ def test_oversized_images_downscaled_and_highres():
     pdf.add_page()
     pdf.image(IMAGE_PATH, w=50)
     pdf.image(IMAGE_PATH, w=pdf.epw)
-    assert len(pdf.images) == 1, pdf.images.keys()
+    assert len(pdf.image_cache.images) == 1, pdf.image_cache.images.keys()
     # Not calling assert_pdf_equal to avoid storing a large binary (1.4M) in this git repo
 
 
@@ -64,7 +64,7 @@ def test_oversized_images_highres_and_downscaled():
     pdf.add_page()
     pdf.image(IMAGE_PATH, w=pdf.epw)
     pdf.image(IMAGE_PATH, w=50)
-    assert len(pdf.images) == 1, pdf.images.keys()
+    assert len(pdf.image_cache.images) == 1, pdf.image_cache.images.keys()
     # Not calling assert_pdf_equal to avoid storing a large binary (1.4M) in this git repo
 
 
@@ -75,9 +75,9 @@ def test_oversized_images_downscale_biggest_1st(tmp_path):
     pdf.add_page()
     pdf.image(IMAGE_PATH, w=50)
     pdf.image(IMAGE_PATH, w=30)
-    assert len(pdf.images) == 2, pdf.images.keys()
+    assert len(pdf.image_cache.images) == 2, pdf.image_cache.images.keys()
     in_use_img_names = _in_use_img_names(pdf)
-    assert len(in_use_img_names) == 1, pdf.images.keys()
+    assert len(in_use_img_names) == 1, pdf.image_cache.images.keys()
     assert_pdf_equal(pdf, HERE / "oversized_images_downscale_biggest_1st.pdf", tmp_path)
 
 
@@ -90,11 +90,11 @@ def test_oversized_images_downscale_biggest_2nd(caplog, tmp_path):
     pdf.image(IMAGE_PATH, w=30)
     pdf.image(IMAGE_PATH, w=50)
     assert "OVERSIZED: Updated low-res image" in caplog.text
-    assert len(pdf.images) == 2, pdf.images.keys()
+    assert len(pdf.image_cache.images) == 2, pdf.image_cache.images.keys()
     in_use_img_names = _in_use_img_names(pdf)
-    assert len(in_use_img_names) == 1, pdf.images.keys()
+    assert len(in_use_img_names) == 1, pdf.image_cache.images.keys()
     assert_pdf_equal(pdf, HERE / "oversized_images_downscale_biggest_2nd.pdf", tmp_path)
 
 
 def _in_use_img_names(pdf):
-    return [name for name, img in pdf.images.items() if img["usages"]]
+    return [name for name, img in pdf.image_cache.images.items() if img["usages"]]